requirements.txt
openai
Flask
python-dotenv
TTS
pydub
scipy
gpt4all
stftpitchshift
gevent
redis
datetime
uuid